Factor out OAuth2ProviderHelper
