+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/core/AuthController.scala
@@ -310,8 +310,7 @@ class AuthController @Inject() (
   }
 
   def signup(provider: String) = Action.async(parse.anyContent) { implicit request =>
-    val authRes = timing(s"[signup($provider)] authenticate") { ProviderController.authenticate(provider) }
-
+    val authRes = ProviderController.authenticate(provider)
     authRes(request).map { result =>
       authHelper.authHandler(request, result) { (_, sess: Session) =>
         // TODO: set FORTYTWO_USER_ID instead of clearing it and then setting it on the next request?
